third_party/WebKit/LayoutTests/media/controls-timeline-expected.txt

EVENT(loadedmetadata)
EXPECTED (video.currentTime == '0') OK
EXPECTED (video.seeking == 'false') OK
EXPECTED (video.currentTime / video.duration > '0.4') OK
EXPECTED (video.currentTime / video.duration < '0.6') OK
EXPECTED (video.seeking == 'true') OK
END OF TEST
